Phase 4 simplify pass (interim — efficiency + reuse)
Efficiency agent: wire_drawing.js mousemove was calling findPortAt
every rAF tick to populate cosmetic metadata (~9000 boundingBox calls
per drag, no observable benefit). Removed per-frame hit-test;
mousemove now does just the SVG path geometry update.

Reuse / quality agent in-flight changes: integration tests/conftest
extracted to share fixtures across the new inspector tests.

231 passed, 1 skipped.
