+++ tests/integration/gui/builder/test_inspector_wire_card.py
@@ -20,8 +20,6 @@ from __future__ import annotations
 
 import json
 from pathlib import Path
-from typing import Any, Iterable, List
-
 
 from phenotypic.gui.builder import _ids as ids
 from phenotypic.gui.builder._layout import (
@@ -37,72 +35,14 @@ from phenotypic.gui.builder._state import (
     state_from_json,
 )
 
+# Component-tree walking helpers shared with ``test_inspector_aux_section.py``;
+# see ``conftest.py`` in this directory.
+from .conftest import _collect_text, _find_by_id, _find_by_type_key
+
 
 FIXTURE_DIR = (
     Path(__file__).resolve().parents[4] / "tests" / "fixtures" / "builder_dag"
 )
-
-
-def _walk(component: Any) -> Iterable[Any]:
-    """Yield *component* and every descendant component recursively.
-
-    Mirrors the ``_walk_components`` helper in ``test_doc_section.py`` —
-    Dash components carry their children on ``.children`` (a single
-    component, a list, or a primitive); the walk descends only into
-    component-shaped entries so primitives don't blow up the iteration.
-    """
-
-    yield component
-    children = getattr(component, "children", None)
-    if children is None:
-        return
-    if isinstance(children, (list, tuple)):
-        for child in children:
-            if child is None or isinstance(child, (str, int, float, bool)):
-                continue
-            yield from _walk(child)
-    elif not isinstance(children, (str, int, float, bool)):
-        yield from _walk(children)
-
-
-def _find_by_id(component: Any, target_id: Any) -> List[Any]:
-    """Return every descendant whose ``id`` equals *target_id*."""
-
-    hits: List[Any] = []
-    for node in _walk(component):
-        node_id = getattr(node, "id", None)
-        if isinstance(target_id, dict):
-            if isinstance(node_id, dict) and node_id == target_id:
-                hits.append(node)
-        elif node_id == target_id:
-            hits.append(node)
-    return hits
-
-
-def _find_by_type_key(component: Any, type_key: str) -> List[Any]:
-    """Return every descendant whose dict-shaped id carries ``type==type_key``."""
-
-    hits: List[Any] = []
-    for node in _walk(component):
-        node_id = getattr(node, "id", None)
-        if isinstance(node_id, dict) and node_id.get("type") == type_key:
-            hits.append(node)
-    return hits
-
-
-def _collect_text(component: Any) -> str:
-    """Flatten every string child / leaf into a single space-joined string."""
-
-    parts: List[str] = []
-    for node in _walk(component):
-        children = getattr(node, "children", None)
-        if isinstance(children, str):
-            parts.append(children)
-        elif isinstance(children, (list, tuple)):
-            for child in children:
-                if isinstance(child, str):
-                    parts.append(child)
-    return " ".join(parts)
 
 
 # ---------------------------------------------------------------------------
